+++ R/map.R
@@ -62,12 +62,9 @@
 #' they are of the same type to avoid coercion. Note: errors if an
 #' \sQuote{errorValue} has been returned or results are of differing type.
 #'
-#' \code{x[.progress]} collects map results whilst showing a simple text
-#' progress indicator of parts completed of the total.
-#'
-#' \code{x[.progress_cli]} collects map results whilst showing a progress bar
-#' from the \CRANpkg{cli} package, if available, with completion percentage and
-#' ETA.
+#' \code{x[.progress]} collects map results whilst showing a progress bar from
+#' the \CRANpkg{cli} package, if installed, with completion percentage and ETA,
+#' or else a simple text progress indicator.
 #'
 #' \code{x[.stop]} collects map results applying early stopping, which stops at
 #' the first failure and cancels remaining operations. Note: operations already
@@ -258,9 +255,6 @@ print.mirai_map <- function(x, ...) {
   )
 )
 
-#' @rdname dot-flat
-#' @export
-#'
 .progress_cli <- compiler::compile(
   quote(
     if (i == 0L) cli::cli_progress_bar(type = NULL, total = xlen, auto_terminate = TRUE, .envir = .) else
@@ -273,6 +267,18 @@ print.mirai_map <- function(x, ...) {
 #'
 .stop <- compiler::compile(
   quote(if (is_error_value(xi)) { stop_mirai(x); stop(sprintf("In index %d:\n%s", i, xi), call. = FALSE) })
+)
+
+.stop_cli <- compiler::compile(
+  quote(
+    if (is_error_value(xi)) {
+      stop_mirai(x)
+      cli::cli_abort(
+        c(i = "In index {i}.", x = xi),
+        call = quote(mirai::mirai_map())
+      )
+    }
+  )
 )
 
 # internals --------------------------------------------------------------------