+++ R/mirai-package.R
@@ -74,6 +74,11 @@
     }
   )
 
+  if (requireNamespace("cli", quietly = TRUE)) {
+    .progress <<- .progress_cli
+    .stop <<- .stop_cli
+  }
+
 }
 
 # nocov end